@@ -1,48 +1,54 @@
 package com.app.delock.delock_application.dashboard;
 
+import android.annotation.SuppressLint;
 import android.content.Intent;
 import android.os.Bundle;
 import android.support.design.widget.BottomNavigationView;
 import android.support.v7.app.AppCompatActivity;
 import android.support.v7.widget.Toolbar;
 import android.view.MenuItem;
+import android.widget.TextView;
+
+import io.ipfs.kotlin.IPFS;
+import io.ipfs.kotlin.IPFSConnection;
+import okhttp3.OkHttpClient;
+
 
 import com.app.delock.delock_application.R;
 import com.app.delock.delock_application.browse.BrowseActivity;
 import com.app.delock.delock_application.my_profile.MyProfileActivity;
 import com.google.android.gms.common.data.DataHolder;
+import com.squareup.moshi.Moshi;
 
 public class DashboardActivity extends AppCompatActivity {
 
+    @SuppressLint("SetTextI18n")
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_dashboard);
-        Toolbar toolbar = (Toolbar) findViewById(R.id.toolbar);
+        Toolbar toolbar = findViewById(R.id.toolbar);
         setSupportActionBar(toolbar);
 
         BottomNavigationView navMenu = findViewById(R.id.navigation);
         navMenu.setSelectedItemId(R.id.navigation_dashboard);
         navMenu.setOnNavigationItemSelectedListener(
-                new BottomNavigationView.OnNavigationItemSelectedListener() {
-                    @Override
-                    public boolean onNavigationItemSelected(MenuItem menuItem) {
-                        Intent intent;
-                        // set item as selected to persist highlight
-                        switch (menuItem.getItemId()) {
-                            case R.id.navigation_home:
-                                intent = new Intent(DashboardActivity.this, BrowseActivity.class);
-                                startActivity(intent);
-                                break;
-                            case R.id.navigation_dashboard:
-                                break;
-                            case R.id.navigation_myProfile:
-                                intent = new Intent(DashboardActivity.this, MyProfileActivity.class);
-                                startActivity(intent);
-                                break;
-                        }
-                        return true;
+                menuItem -> {
+                    Intent intent;
+                    // set item as selected to persist highlight
+                    switch (menuItem.getItemId()) {
+                        case R.id.navigation_home:
+                            intent = new Intent(DashboardActivity.this, BrowseActivity.class);
+                            startActivity(intent);
+                            break;
+                        case R.id.navigation_dashboard:
+                            break;
+                        case R.id.navigation_myProfile:
+                            intent = new Intent(DashboardActivity.this, MyProfileActivity.class);
+                            startActivity(intent);
+                            break;
                     }
+                    return true;
                 });
     }
 
